Moved classes up - they're reused now (#186)
