+++ dz3-swing/src/main/java/net/sf/dz3/view/swing/sensor/SensorPanel.java
@@ -1,8 +1,8 @@
 package net.sf.dz3.view.swing.sensor;
 
 import net.sf.dz3.device.sensor.AnalogSensor;
+import net.sf.dz3.view.swing.EntityPanel;
 import net.sf.dz3.view.swing.ScreenDescriptor;
-import net.sf.dz3.view.swing.thermostat.EntityPanel;
 
 import javax.swing.JPanel;
 import java.awt.event.KeyEvent;
